endomondo import tweaks added time and option to clear or update workouts

+++ app/Console/Commands/ImportEndomondo.php
@@ -17,7 +17,7 @@ class ImportEndomondo extends Command
      *
      * @var string
      */
-    protected $signature = 'import:endomondo';
+    protected $signature = 'import:endomondo {--clear}';
 
     /**
      * The console command description.
@@ -64,20 +64,31 @@ class ImportEndomondo extends Command
         }
 
         $data['workouts'] = array_reverse($data['workouts']);
-        $this->info( 'Deleting workouts...' );
-        DB::table('workouts')->where('user_id', '=', $user->id)->delete();
+
+        if( $this->option('clear') ) {
+            $this->info('Deleting workouts...');
+            DB::table('workouts')->where('user_id', '=', $user->id)->delete();
+        }
+
+        $workoutExternalIds = Workout::withoutGlobalScope('user_id')->whereNotNull( 'external_id' )->pluck('id', 'external_id');
 
         $this->info( 'Importing workouts...' );
 
+        $importCount = 0;
         $bar = $this->output->createProgressBar(count($data['workouts']));
-
         foreach( $data['workouts'] as $index => $endomondoWorkout ){
 
             $type = $endomondoWorkout->getTypeId();
-            if( isset( $typeMap[$type] ) ) {
+            $externalId = $endomondoWorkout->getId();
+
+            if( isset( $typeMap[$type] ) && empty( $workoutExternalIds[$externalId] ) ) {
+
+                $importCount++;
 
                 $workout = [
+                    'external_id' => $externalId,
                     'title' => 'Endomondo workout',
+                    'time' => $endomondoWorkout->getStart(),
                     'type' => $typeMap[$type],
                     'import_filename' => '',
                     'user_id' => $user->id,
@@ -108,6 +119,6 @@ class ImportEndomondo extends Command
         $bar->finish();
         $this->output->newLine();
 
-        $this->info(  sprintf("Imported %s workouts from Endomondo", count($data['workouts']) ) );
+        $this->info(  sprintf("Imported %s new workouts from Endomondo, skipped the existing: %s", $importCount, count($data['workouts']) - $importCount ) );
     }
 }
